refactor(package.controller): controller 중 domain 관련 controller를 하나로 묶었다.
