+++ README.md
@@ -62,7 +62,7 @@
     - 예외
         - [X] 노선이 없는 경우
         - [X] 노선에 역이 없는 경우
-        - [ ] 노선에 포함된 역이 두 개 이하일 때는 역을 제거할 수 없다.
+        - [x] 노선에 포함된 역이 두 개 이하일 때는 역을 제거할 수 없다.
 
 ### 지하철 노선에 등록된 역 조회 기능
 - [X] 노선의 상행 종점부터 하행 종점까지 연결된 순서대로 역 목록을 조회할 수 있다.
